@@ -19,24 +19,13 @@
     "lint:eslint": "eslint --ext .js,.vue,.ts --fix .",
     "lint:prettier": "prettier --check --write .",
     "lint:stylelint": "stylelint packages/*/src/**/*.scss --fix",
-    "prepare": "husky install",
     "server:dev": "vercel dev ./example --listen 9090",
     "test": "vitest"
   },
   "devDependencies": {
-    "typescript": "^2.5.3",
+    "@changesets/cli": "^2.24.1",
+    "typescript": "^4.7.4",
     "vercel": "27.2.0"
   },
-  "private": "true",
-  "pnpm": {
-    "peerDependencyRules": {
-      "ignoreMissing": [
-        "@algolia/client-search",
-        "@types/react",
-        "react",
-        "react-dom",
-        "typescript"
-      ]
-    }
-  }
+  "private": "true"
 }
