@@ -1,7 +1,5 @@
-import Application from "@dragondyt/sakura-server";
-export default Application({
-    // @ts-ignore
-    async postSave(comment) {
-        // do what ever you want after save comment
+import Route from "@dragondyt/sakura-server/lib";
+export default Route({
+    async postSave() {
     },
 })
