@@ -1,21 +1,15 @@
-export default function() {
+export default function dashboard() {
     return (ctx: ThinkContext) => {
         ctx.type = 'html';
         ctx.body = `<!doctype html>
-<html lang="zh-CN">
+<html>
   <head>
     <meta charset="utf-8">
     <title>Waline Management System</title>
     <meta name="viewport" content="width=device-width,initial-scale=1">
   </head>
   <body>
-    <script>
-    window.SITE_URL = ${JSON.stringify(process.env.SITE_URL)};
-    window.SITE_NAME = ${JSON.stringify(process.env.SITE_NAME)};
-    </script>
-    <script src="${
-            process.env.WALINE_ADMIN_MODULE_ASSET_URL || '//unpkg.com/@waline/admin'
-        }"></script>
+  fack
   </body>
 </html>`;
     };
